pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cluster-dispatch"
version = "0.1.0"
description = "Track and run analysis projects on remote compute servers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
cdp = "cluster_dispatch.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
